no longer hard-code output of test_usage test

+++ tools/test_usage.csh
@@ -3,6 +3,16 @@
 setenv TERM xterm
 setenv PYRAF_NO_DISPLAY 1
 
+# First create expected output dynamically from the pyraf script itself
+# (use `which pyraf` to get the pyraf script to be used below)
+/bin/rm -f test_usage.expected >& /dev/null
+if ($#argv != 1) then
+   echo "usage: $0 full-path-to-pyraf-source-script"
+   exit 1
+endif
+set pyrafscript = $argv[1]
+cat $pyrafscript |sed -n '3,$ p' |sed '/"""/,$ d' > test_usage.expected
+
 #
 # TRY '--help'
 /bin/rm -f test_usage1.txt >& /dev/null
@@ -19,7 +29,7 @@ endif
 cat test_usage1.txt.orig | sed -n '/Copyright/,$ p' > test_usage1.txt
 
 # Do the diff
-/usr/bin/diff -Bw test_usage.save test_usage1.txt
+/usr/bin/diff -Bw test_usage.expected test_usage1.txt
 if ($status != 0) then
    echo ERROR DURING DIFF 1
    exit 1
@@ -34,7 +44,7 @@ pyraf --silent -h >& test_usage2.txt
 cat test_usage2.txt.orig | sed -n '/Copyright/,$ p' > test_usage2.txt
 
 # Do the diff
-/usr/bin/diff -Bw test_usage.save test_usage2.txt
+/usr/bin/diff -Bw test_usage.expected test_usage2.txt
 if ($status != 0) then
    echo ERROR DURING DIFF 2
    exit 1
